+++ apps/decodex/src/orchestrator/status/review_orchestration/marker.rs
@@ -1,0 +1,190 @@
+use crate::{
+	orchestrator::status::{
+		self, PostReviewLaneClassification, PostReviewLaneSnapshot, PostReviewRuntimeState,
+		PrivateExecutionEvent, PullRequestReviewState, ReviewCheckpointArtifactLookup,
+		ReviewOrchestrationMarker, Value,
+	},
+	prelude::Result,
+};
+
+pub(in crate::orchestrator) fn load_post_review_orchestration_marker(
+	snapshot: &PostReviewLaneSnapshot,
+	review_state: &PullRequestReviewState,
+	classification: &mut PostReviewLaneClassification,
+	runtime_state: Option<PostReviewRuntimeState<'_>>,
+) -> Result<Option<ReviewOrchestrationMarker>> {
+	let review_handoff = snapshot
+		.review_handoff
+		.as_ref()
+		.expect("review handoff should exist before orchestration classification");
+	let orchestration_marker = if let Some(runtime_state) = runtime_state {
+		runtime_state.state_store.review_orchestration_marker(
+			runtime_state.project_id,
+			&snapshot.issue.id,
+			review_handoff,
+		)?
+	} else {
+		None
+	};
+	let Some(orchestration_marker) = orchestration_marker else {
+		if clean_current_head_review_repair_writeback_pending(
+			snapshot,
+			review_state,
+			runtime_state,
+		)? {
+			classification.reason =
+				String::from("review_repair_writeback_missing_lifecycle_marker");
+			classification.readback_warning =
+				Some(String::from("review_repair_writeback_missing_lifecycle_marker"));
+
+			return Ok(None);
+		}
+
+		classification.reason = String::from("external_review_request_pending");
+
+		return Ok(None);
+	};
+
+	if let Some(reason) =
+		validate_review_orchestration_marker(snapshot, review_state, &orchestration_marker)
+	{
+		if reason == "review_orchestration_head_mismatch"
+			&& clean_current_head_review_repair_writeback_pending(
+				snapshot,
+				review_state,
+				runtime_state,
+			)? {
+			classification.reason = String::from("review_repair_writeback_stale_lifecycle_marker");
+			classification.readback_warning =
+				Some(String::from("review_repair_writeback_stale_lifecycle_marker"));
+
+			return Ok(None);
+		}
+
+		*classification = status::blocked_post_review_lane_from_state(review_state, reason);
+
+		return Ok(None);
+	}
+
+	Ok(Some(orchestration_marker))
+}
+
+pub(in crate::orchestrator) fn clean_current_head_review_repair_writeback_pending(
+	snapshot: &PostReviewLaneSnapshot,
+	review_state: &PullRequestReviewState,
+	runtime_state: Option<PostReviewRuntimeState<'_>>,
+) -> Result<bool> {
+	let Some(runtime_state) = runtime_state else {
+		return Ok(false);
+	};
+	let Some(local_head_oid) = snapshot.local_head_oid.as_deref() else {
+		return Ok(false);
+	};
+
+	if review_state.head_ref_oid != local_head_oid
+		|| review_state.head_ref_name != snapshot.worktree.branch_name()
+	{
+		return Ok(false);
+	}
+
+	let events = runtime_state
+		.state_store
+		.list_private_execution_events_for_issue(runtime_state.project_id, &snapshot.issue.id)?;
+
+	for terminal_event in events.iter().rev() {
+		if !review_repair_terminal_finalize_event_matches_snapshot(terminal_event, snapshot) {
+			continue;
+		}
+
+		let Some(intent_event) = events.iter().rev().find(|event| {
+			event.run_id() == terminal_event.run_id()
+				&& event.attempt_number() == terminal_event.attempt_number()
+				&& review_repair_completion_intent_matches_current_head(
+					event,
+					snapshot,
+					review_state,
+					local_head_oid,
+				)
+		}) else {
+			continue;
+		};
+		let Some(checkpoint) = runtime_state.state_store.review_checkpoint_artifact(
+			ReviewCheckpointArtifactLookup {
+				project_id: runtime_state.project_id,
+				issue_id: &snapshot.issue.id,
+				phase: "repair",
+				review_level: runtime_state.review_level.as_str(),
+				head_sha: local_head_oid,
+			},
+		)?
+		else {
+			continue;
+		};
+
+		if checkpoint.status() == "clean"
+			&& checkpoint.head_sha() == local_head_oid
+			&& checkpoint.run_id() == intent_event.run_id()
+			&& checkpoint.attempt_number() == intent_event.attempt_number()
+		{
+			return Ok(true);
+		}
+	}
+
+	Ok(false)
+}
+
+pub(in crate::orchestrator) fn review_repair_terminal_finalize_event_matches_snapshot(
+	event: &PrivateExecutionEvent,
+	snapshot: &PostReviewLaneSnapshot,
+) -> bool {
+	let payload = event.payload();
+
+	event.event_type() == "terminal_finalize"
+		&& payload.get("path").and_then(Value::as_str) == Some("review_repair")
+		&& payload.get("mode").and_then(Value::as_str) == Some("repair")
+		&& payload.get("branch").and_then(Value::as_str) == Some(snapshot.worktree.branch_name())
+		&& payload.get("worktree_path").and_then(Value::as_str)
+			== Some(snapshot.worktree.worktree_path().display().to_string().as_str())
+}
+
+pub(in crate::orchestrator) fn review_repair_completion_intent_matches_current_head(
+	event: &PrivateExecutionEvent,
+	snapshot: &PostReviewLaneSnapshot,
+	review_state: &PullRequestReviewState,
+	local_head_oid: &str,
+) -> bool {
+	let payload = event.payload();
+
+	event.event_type() == "review_completion_intent"
+		&& payload.get("path").and_then(Value::as_str) == Some("review_repair")
+		&& payload.get("mode").and_then(Value::as_str) == Some("repair")
+		&& payload.get("branch").and_then(Value::as_str) == Some(snapshot.worktree.branch_name())
+		&& payload.get("worktree_path").and_then(Value::as_str)
+			== Some(snapshot.worktree.worktree_path().display().to_string().as_str())
+		&& payload.get("pr_url").and_then(Value::as_str) == Some(review_state.url.as_str())
+		&& payload.get("pr_head_ref").and_then(Value::as_str)
+			== Some(review_state.head_ref_name.as_str())
+		&& payload.get("pr_head_oid").and_then(Value::as_str) == Some(local_head_oid)
+}
+
+pub(in crate::orchestrator) fn validate_review_orchestration_marker(
+	snapshot: &PostReviewLaneSnapshot,
+	review_state: &PullRequestReviewState,
+	marker: &ReviewOrchestrationMarker,
+) -> Option<&'static str> {
+	let Some(local_head_oid) = snapshot.local_head_oid.as_deref() else {
+		return Some("worktree_head_missing");
+	};
+
+	if marker.branch_name() != snapshot.worktree.branch_name() {
+		return Some("review_orchestration_branch_mismatch");
+	}
+	if marker.pr_url() != review_state.url {
+		return Some("review_orchestration_pr_mismatch");
+	}
+	if marker.head_sha() != local_head_oid {
+		return Some("review_orchestration_head_mismatch");
+	}
+
+	None
+}
